requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.38.8
botocore==1.38.8
Django==5.2
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.9.0
